Remove feature flag for new HeaderBLock and remove the deprecated HeaderBlock
