+++ apps/store/src/app/[locale]/StoryblokLayout.tsx
@@ -2,10 +2,8 @@ import { clsx } from 'clsx'
 import type { ReactNode } from 'react'
 import { FooterBlock } from '@/blocks/FooterBlock/FooterBlock'
 import { HeaderBlock } from '@/blocks/HeaderBlock/HeaderBlock'
-import { HeaderBlock as HeaderBlockNew } from '@/blocks/HeaderBlockNew/HeaderBlock'
 import type { GlobalStory } from '@/services/storyblok/storyblok'
 import { filterByBlockType } from '@/services/storyblok/Storyblok.helpers'
-import { Features } from '@/utils/Features'
 import { wrapper } from './StoryblokLayout.css'
 
 export const StoryblokLayout = ({
@@ -22,13 +20,9 @@ export const StoryblokLayout = ({
 
   return (
     <div className={clsx(wrapper, className)}>
-      {headerBlock.map((nestedBlock) =>
-        Features.enabled('NEW_HEADER') ? (
-          <HeaderBlockNew key={nestedBlock._uid} blok={nestedBlock} />
-        ) : (
-          <HeaderBlock key={nestedBlock._uid} blok={nestedBlock} />
-        ),
-      )}
+      {headerBlock.map((nestedBlock) => (
+        <HeaderBlock key={nestedBlock._uid} blok={nestedBlock} />
+      ))}
       {children}
       {footerBlock.map((nestedBlock) => (
         <FooterBlock key={nestedBlock._uid} blok={nestedBlock} />
