+++ apps/store/src/blocks/HeaderBlock/ProductMenuBlock.tsx
@@ -5,7 +5,7 @@ import clsx from 'clsx'
 import { useTranslation } from 'next-i18next'
 import { sprinkles } from 'ui/src/theme/sprinkles.css'
 import type { ButtonBlockProps } from '@/blocks/ButtonBlock'
-import { MenuItemBlock, type MenuItemBlockProps } from '@/blocks/HeaderBlockNew/MenuItemBlock'
+import { MenuItemBlock, type MenuItemBlockProps } from '@/blocks/HeaderBlock/MenuItemBlock'
 import { ButtonNextLink } from '@/components/ButtonNextLink'
 import {
   navigationItem,
