+++ apps/store/src/blocks/HeaderBlock/SubMenuBlock.tsx
@@ -1,43 +1,68 @@
-'use client'
 import * as NavigationMenuPrimitive from '@radix-ui/react-navigation-menu'
 import { storyblokEditable } from '@storyblok/react'
 import clsx from 'clsx'
-import { MenuItemBlock, type MenuItemBlockProps } from '@/blocks/HeaderBlockNew/MenuItemBlock'
+import { MinusIcon, PlusIcon } from 'ui'
+import { MenuItemBlock, type MenuItemBlockProps } from '@/blocks/HeaderBlock/MenuItemBlock'
+import { ButtonNextLink } from '@/components/ButtonNextLink'
 import {
   navigationItem,
+  navigationItemSubMenu,
+  navigationItemSupportMenu,
   navigationMenuWrapper,
   navigationSecondaryItem,
   navigationSecondaryList,
-  navigationItemGeneralMenu,
-  navigationItemSubMenu,
 } from '@/components/HeaderNew/Header.css'
 import { NavigationContent } from '@/components/HeaderNew/NavigationContent'
 import { SecondaryNavigationLink } from '@/components/HeaderNew/NavigationLink/NavigationLink'
-import { NavigationTriggerGeneralMenu } from '@/components/HeaderNew/NavigationTrigger/NavigationTriggerGeneralMenu'
+import { NavigationTrigger } from '@/components/HeaderNew/NavigationTrigger/NavigationTrigger'
+import { openIcon, closeIcon } from '@/components/HeaderNew/NavigationTrigger/NavigationTrigger.css'
 import type { ExpectedBlockType, SbBaseBlockProps } from '@/services/storyblok/storyblok'
 import { filterByBlockType, getLinkFieldURL } from '@/services/storyblok/Storyblok.helpers'
 
-export type GeneralMenuBlockProps = SbBaseBlockProps<{
+export type SubMenuBlockProps = SbBaseBlockProps<{
   name: string
   menuItems: ExpectedBlockType<MenuItemBlockProps>
   currentActiveItem?: string
-}>
-export const GeneralMenuBlock = ({ blok }: GeneralMenuBlockProps) => {
+}> & { variant: 'mobile' | 'desktop' }
+export const SubMenuBlock = ({ blok, variant }: SubMenuBlockProps) => {
   const filteredMenuItems = filterByBlockType(blok.menuItems, MenuItemBlock.blockName)
 
   if (filteredMenuItems.length === 0) {
     return null
   }
 
+  const firstMenuItem = filteredMenuItems[0]
+
   return (
     <NavigationMenuPrimitive.Item
-      className={clsx(navigationItem, navigationItemGeneralMenu, navigationItemSubMenu)}
+      className={clsx(navigationItem, navigationItemSubMenu, navigationItemSupportMenu)}
       value={blok.name}
       {...storyblokEditable(blok)}
     >
-      <NavigationTriggerGeneralMenu>{blok.name}</NavigationTriggerGeneralMenu>
-      {/* Prevent closing the menu when cursor leaves content */}
-      <NavigationContent onPointerLeave={(event) => event.preventDefault()}>
+      {variant === 'mobile' ? (
+        <NavigationTrigger
+          onPointerEnter={(event) => event.preventDefault()}
+          onPointerLeave={(event) => event.preventDefault()}
+          onPointerMove={(event) => event.preventDefault()}
+        >
+          <button>
+            {blok.name}
+            <PlusIcon className={openIcon} size="1rem" />
+            <MinusIcon className={closeIcon} size="1rem" />
+          </button>
+        </NavigationTrigger>
+      ) : (
+        <NavigationTrigger>
+          <ButtonNextLink
+            size="medium"
+            variant="secondary"
+            href={getLinkFieldURL(firstMenuItem.link, firstMenuItem.name)}
+          >
+            {blok.name}
+          </ButtonNextLink>
+        </NavigationTrigger>
+      )}
+      <NavigationContent>
         <div className={navigationMenuWrapper}>
           <NavigationMenuPrimitive.Sub defaultValue={blok.name}>
             <NavigationMenuPrimitive.List className={navigationSecondaryList}>
@@ -62,4 +87,4 @@ export const GeneralMenuBlock = ({ blok }: GeneralMenuBlockProps) => {
     </NavigationMenuPrimitive.Item>
   )
 }
-GeneralMenuBlock.blockName = 'generalMenu'
+SubMenuBlock.blockName = 'subMenu'
