+++ apps/store/src/components/LayoutWithMenu/LayoutWithMenu.tsx
@@ -2,11 +2,9 @@ import { clsx } from 'clsx'
 import { type ReactElement } from 'react'
 import { FooterBlock } from '@/blocks/FooterBlock/FooterBlock'
 import { HeaderBlock } from '@/blocks/HeaderBlock/HeaderBlock'
-import { HeaderBlock as HeaderBlockNew } from '@/blocks/HeaderBlockNew/HeaderBlock'
 import { wrapper } from '@/components/LayoutWithMenu/LayoutWithMenu.css'
 import type { GlobalStory } from '@/services/storyblok/storyblok'
 import { filterByBlockType } from '@/services/storyblok/Storyblok.helpers'
-import { Features } from '@/utils/Features'
 import type { GlobalProductMetadata } from './fetchProductMetadata'
 import { GLOBAL_PRODUCT_METADATA_PROP_NAME } from './fetchProductMetadata'
 import { useHydrateProductMetadata } from './productMetadataHooks'
@@ -36,13 +34,9 @@ export const LayoutWithMenu = (props: LayoutWithMenuProps) => {
     <>
       <div className={clsx(wrapper, className)}>
         {!props.hideMenu &&
-          headerBlock.map((nestedBlock) =>
-            Features.enabled('NEW_HEADER') ? (
-              <HeaderBlockNew key={nestedBlock._uid} blok={nestedBlock} />
-            ) : (
-              <HeaderBlock key={nestedBlock._uid} blok={nestedBlock} />
-            ),
-          )}
+          headerBlock.map((nestedBlock) => (
+            <HeaderBlock key={nestedBlock._uid} blok={nestedBlock} />
+          ))}
         {props.children}
         {!props.hideFooter &&
           footerBlock.map((nestedBlock) => (
